+++ all/native/styles/TextStyle.h
@@ -61,7 +61,7 @@ namespace carto {
                   float renderScale,
                   const std::string& fontName,
                   const std::string& textField,
-                  int fontSize,
+                  float fontSize,
                   const Color& strokeColor,
                   float strokeWidth);
         virtual ~TextStyle();
@@ -88,7 +88,7 @@ namespace carto {
          * Returns the font's size.
          * @return The size of the font in points.
          */
-        int getFontSize() const;
+        float getFontSize() const;
 
         /**
          * Returns the color of the stroke.
@@ -109,7 +109,7 @@ namespace carto {
         
         std::string _textField;
 
-        int _fontSize;
+        float _fontSize;
 
         Color _strokeColor;
